rewrite curried function calls to traditional style
Many people do not understand or have issues with the current style of
function declaration (expressions in this case). Another issues is that
the user will not get any warnings/errors if not all parameter
(functions) are provided. For these reasons we rewrite the curried
function expressions to traditional function declarations. For cases,
where a curried function is needed, the function needs to be binded.