add default function for fast setup
Introduce a default function for a fast setup of forms/inputs.
Most users use the simple form of patching. Instead of doing this task
manually, it is a good idea to have a setup function that already takes care of
this task. This makes it easier for new users to get started.