adjust editorconfig to avoid interference
Because `trim_trailing_whitespace` was applied to all file types, it
also interfered with git manual hunk edits, which did not apply
correctly, when the last context line was only an empty line (space).
To avoid that we will not default to `trim_trailing_whitespace` and only
apply it on file types that are actually used in the project.