run validator on listed elements
Previously, only submittable elements could be passed to the validator.
This is not quite correct because all listed elements are candidates for
constraint validation.

https://dev.w3.org/html5/spec-LC/association-of-controls-and-forms.html#candidate-for-constraint-validation

Additionally, support for EventListener has been removed because the
type cannot be guaranteed by the event target. Instead, the element of
the correct type must now be passed directly.