reset validity state of inputs (barred or not)
If an input is invalid and subsequently deactivated, then any revalidation
cannot reset the validity. As a result, processes that check for validity
would not work correctly because the input has an incorrect state.

Resetting the validity of an input should always work, regardless of whether
the input is excluded from validation or not.